requirements.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.108.0
h11==0.14.0
idna==3.6
isodate==0.6.1
lxml==4.9.3
onvif-zeep==0.2.12
platformdirs==4.1.0
pydantic==2.5.3
pydantic-core==2.14.6
pytz==2023.3.post1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.32.0.post1
typing-extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
zeep==4.2.1